fixup! Detect and prevent recursive config parsing
