+++ test/recipes/30-test_prov_config.t
@@ -20,13 +20,11 @@ use lib bldtop_dir('.');
 
 my $no_fips = disabled('fips') || ($ENV{NO_FIPS} // 0);
 
-plan tests => 3;
+plan tests => 2;
 
-ok(run(test(["prov_config_test", srctop_file("test", "default.cnf")])),
+ok(run(test(["prov_config_test", srctop_file("test", "default.cnf"),
+                                 srctop_file("test", "recursive.cnf")])),
     "running prov_config_test default.cnf");
-
-ok(run(test(["prov_recurse_config_test", srctop_file("test", "recursive.cnf")])),
-    "running prov_recurse_config_test");
 
 SKIP: {
     skip "Skipping FIPS test in this build", 1 if $no_fips;
